fix(npc): fix cycles display

+++ npc/include/util/iringbuf.hpp
@@ -7,7 +7,8 @@
 #include "mem/paddr.hpp"
 #include "util/ringbuf.hpp"
 
-using IRingBuf = baudvine::RingBuf<std::tuple<paddr_t, uint32_t, uint8_t>, CONFIG_IRINGBUF_NR_ELEM>;
+using IRingBuf =
+    baudvine::RingBuf<std::tuple<paddr_t, uint32_t, uint16_t>, CONFIG_IRINGBUF_NR_ELEM>;
 
 extern IRingBuf iringbuf;
